+++ lizard_progress/views/download.py
@@ -402,14 +402,9 @@ def protected_download_export_run(request, project_slug, export_run_id):
     file_path = export_run.file_path
     logger.debug("File path: " + file_path)
 
-    nginx_path = '/'.join([
-        '/protected', 'export',
-        export_run.activity.project.organization.name,
-        os.path.basename(file_path)])
-
     # This is where the magic takes place.
     response = HttpResponse()
-    response['X-Accel-Redirect'] = nginx_path  # Nginx
+    response['X-Accel-Redirect'] = export_run.nginx_path  # Nginx
 
     # Unset the Content-Type as to allow for the webserver
     # to determine it.